+++ src/integrationTest/java/uk/gov/hmcts/reform/preapi/services/UserAuthenticationServiceIT.java
@@ -10,7 +10,7 @@ import org.springframework.security.authentication.BadCredentialsException;
 import uk.gov.hmcts.reform.preapi.Application;
 import uk.gov.hmcts.reform.preapi.entities.AppAccess;
 import uk.gov.hmcts.reform.preapi.enums.CourtType;
-import uk.gov.hmcts.reform.preapi.security.UserAuthenticationService;
+import uk.gov.hmcts.reform.preapi.security.service.UserAuthenticationService;
 import uk.gov.hmcts.reform.preapi.util.HelperFactory;
 
 import java.util.List;
