+++ src/test/java/uk/gov/hmcts/reform/preapi/controller/RoleControllerTest.java
@@ -10,7 +10,7 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import uk.gov.hmcts.reform.preapi.controllers.RoleController;
 import uk.gov.hmcts.reform.preapi.dto.RoleDTO;
-import uk.gov.hmcts.reform.preapi.security.UserAuthenticationService;
+import uk.gov.hmcts.reform.preapi.security.service.UserAuthenticationService;
 import uk.gov.hmcts.reform.preapi.services.RoleService;
 
 import java.util.List;
